list manipulation
1.a In this question you will write a program that takes two lists of the same length and returns a single list with the items paired together.
example :
# let rec pairlists twolists = ...
val pairlists : 'a list * 'b list -> ('a * 'b) list = <fun>

b. If (wi) is a sequence of weights and (vi) is a sequence of values we define the weighted mean or average by the formula
mean=∑wi⋅vi/∑wi.
In this question you are given two lists of floats as input. The first list is a list of weights and the second is the list of values. Implement a function wmean which calculates the weighted mean of the values with the given weights. 

2. This question involves two simple exercises in list manipulation. In the first one you are given a pair: the first member of the pair is an item and the second member is a list of items of the same type as the first member. The output is a boolean that says whether the item is in the list or not. The second function takes a pair of an item and a list and returns a list with the item removed. If the item is not in the list then the same list is returned.

3. In this question we want you to find the largest item in a list of integers. You can assume that the lists are nonempty. The top level declaration is nonrecursive because there is an internal recursive function.

4. In this question we want you to use find_max and remove to implement a selection sort algorithm. You can assume that the list contains no duplicates.